Move REPL code to its own file
